LLM/Phishing_Detection_System/Phishing_Email_Detection/FastAPI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
psutil==5.9.6
python-email-validator==2.1.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1